+++ wp-content/themes/workingnyc/views/newsletter.twig
@@ -67,8 +67,9 @@
                 </svg>
               </div>
 
-              <div class="c-alert__body">
-                <p data-js-alert="text"></p>
+              <div class="c-alert__body bg-[#C9EEFE]">
+                <b>{{ __('Check your email to confirm your subscription.', 'WNYC') }}</b>
+                <p data-js-alert="text">{{ __('Please click the confirmation link to start receiving our newsletter.', 'WNYC') }}</p>
               </div>
             </article>
           </form>
